backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.0
pydantic==2.9.2
pydantic-settings==2.5.2
pdfplumber==0.11.0
python-docx==1.1.2
python-multipart==0.0.9
weasyprint==62.3
jinja2==3.1.4
slowapi==0.1.9

